+++ Model/Sorghum/Roots.h
@@ -13,43 +13,43 @@ class Roots : public PlantPart
 
 // Parameters ----------------------------------------------------------
 
-   float initialRootDepth;
-   float initialDM;
+   double initialRootDepth;
+   double initialDM;
 
-   vector<float> dLayer;
-   vector<float> xf;
-   float profileDepth;
+   vector<double> dLayer;
+   vector<double> xf;
+   double profileDepth;
    int    nLayers;
 
    TableFn swRoot;
    TableFn rldFn;
 
-   vector<float> rootDepthRate;    // mm/day for each stage
-   float specificRootLength;
+   vector<double> rootDepthRate;    // mm/day for each stage
+   double specificRootLength;
 
    // senescence
-   float dmRootSenFrac;
+   double dmRootSenFrac;
 
    // nitrogen
-   float initialNConc;
-   float targetNConc;
+   double initialNConc;
+   double targetNConc;
 
 
 //  Variables  -----------------------------------------------------
-   float rootDepth;
-   float rootFront;
-   float leftDist;                 // left distance of 2d soil square
-   float rightDist;                // right distance of 2d soil square
-   float dltRootDepth;             // daily increment
-   float dltRootFront;             // daily increment
+   double rootDepth;
+   double rootFront;
+   double leftDist;                 // left distance of 2d soil square
+   double rightDist;                // right distance of 2d soil square
+   double dltRootDepth;             // daily increment
+   double dltRootFront;             // daily increment
    int   currentLayer;             // number of the layer that the roots are in now (starts at 0)
-   float lastLayerPropn;           // proportion of the currentLayer occupied
+   double lastLayerPropn;           // proportion of the currentLayer occupied
 
    // length
-   vector<float> rootLength;
-   vector<float> dltRootLength;
-   vector<float> dltScenescedRootLength;
-   vector<float> rlvFactor;
+   vector<double> rootLength;
+   vector<double> dltRootLength;
+   vector<double> dltScenescedRootLength;
+   vector<double> rlvFactor;
 
 
 // Private Methods -------------------------------------------------------
@@ -57,15 +57,15 @@ class Roots : public PlantPart
    void  doLayerRegistrations(void);
    void  initialize(void);
    void  calcInitialLength(void);
-   float calcDltRootDepth(float stage);
-   float calcDltRootFront(float stage);
-   float swAvailFactor(int layer);
-   float layerProportion(void);
-   float getRootArea(float top, float bottom, float rootLength, float dist);
+   double calcDltRootDepth(double stage);
+   double calcDltRootFront(double stage);
+   double swAvailFactor(int layer);
+   double layerProportion(void);
+   double getRootArea(double top, double bottom, double rootLength, double dist);
 
-   float totalBiomass(void)const{return dmGreen + dmSenesced;}
-   float totalN(void)const{return nGreen + nSenesced;}
-   float totalP(void)const{return pGreen + pSenesced;}
+   double totalBiomass(void)const{return dmGreen + dmSenesced;}
+   double totalN(void)const{return nGreen + nSenesced;}
+   double totalP(void)const{return pGreen + pSenesced;}
 
 
 // public Methods -------------------------------------------------------
@@ -83,14 +83,14 @@ class Roots : public PlantPart
 
    // biomass
    void  calcSenescence(void);
-   void  partitionDM(float dltDM);
+   void  partitionDM(double dltDM);
 
-   float getRootDepth(void)const{return rootDepth;}
-   float getRootFront(void)const{return rootFront;}
-   float RootProportionInLayer(int layer);
+   double getRootDepth(void)const{return rootDepth;}
+   double getRootFront(void)const{return rootFront;}
+   double RootProportionInLayer(int layer);
 
-   float calcNDemand(void);
-   float calcPDemand(void);
+   double calcNDemand(void);
+   double calcPDemand(void);
 
    void incorporateResidue(void);
